Adds code to the image classifier file
this is mostly a straight copy paste from the examples files provided on the DIGITS github page. Work still needs to be done to make the code work nice with classes, since originally it was made to be run a single time.

also deleted 'myFile.py', because it is redundant.