Adds the ability to rename parking spots
while originally each spot was going to have an id number, i have just changed that to be a generic id. any string is a valid name. there is not yet any checking for repeated ids.

changed variable names to reflect this, also rearranged the main app to be more self contained (dont post to root, post to the frame).