adds jes's spot labeling
because the code was so far behind, jeswin's changes were manually brought into the master branch, and his branch will be closed. 

the change brought over is the labeling of spots on the canvas with their id. 

a change that could not be brought over is the scrollbar to go with the list of parking spots. this is beacuse of some implementation issue that resets the scroll to 0 no matter what.