changed the way images are stored internally
instead of saving the path to the image internally and loading it everytime it is needed, the parkinglot class now works with a loaded cv2 image passing around instead of the string. 

This change, in conjuction with the added image_swapper function,  allows a much cleaner system for changing the current image. image_swapper exists on a separate thread and populates a given directory with random images from another given directory.